refactor: Move delete and deactivated account to security controller, and create test for security tests
