+++ app/Http/Controllers/Settings/SecurityController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers\Settings;
 
 use App\Enums\SocialProvidersEnum;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Settings\ProfileUpdateRequest;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -37,30 +36,6 @@ class SecurityController extends Controller
     }
 
     /**
-     * Update the user's profile settings.
-     */
-    public function update(ProfileUpdateRequest $request): RedirectResponse
-    {
-
-        $inputs = $request->validated();
-        $image = $request->file('image');
-        if ($image) {
-            $inputs['image_url'] = $image->store('images', 'public');
-            unset($inputs['image']);
-        }
-
-        $request->user()->fill($inputs);
-
-        if ($request->user()->isDirty('email')) {
-            $request->user()->email_verified_at = null;
-        }
-
-        $request->user()->save();
-
-        return to_route('profile.edit');
-    }
-
-    /**
      * Delete the user's account.
      */
     public function destroy(Request $request): RedirectResponse
@@ -74,6 +49,24 @@ class SecurityController extends Controller
         Auth::logout();
 
         $user->delete();
+
+        $request->session()->invalidate();
+        $request->session()->regenerateToken();
+
+        return redirect('/');
+    }
+
+    public function deactivate(Request $request): RedirectResponse
+    {
+        $request->validate([
+            'password' => ['required', 'current_password'],
+        ]);
+
+        $user = $request->user();
+
+        Auth::logout();
+
+        $user->deactivate();
 
         $request->session()->invalidate();
         $request->session()->regenerateToken();
