+++ routes/settings.php
@@ -11,8 +11,6 @@ Route::middleware('auth')->group(function () {
 
     Route::get('settings/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::post('settings/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('settings/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-    Route::put('settings/profile', [ProfileController::class, 'deactivate'])->name('profile.deactivate');
 
     Route::get('settings/password', [PasswordController::class, 'edit'])->name('password.edit');
     Route::put('settings/password', [PasswordController::class, 'update'])->name('password.update');
@@ -21,5 +19,9 @@ Route::middleware('auth')->group(function () {
         return Inertia::render('settings/appearance');
     })->name('appearance');
 
-    Route::get('settings/security', [SecurityController::class, 'edit'])->name('security.edit');
+    Route::controller(SecurityController::class)->name('security.')->prefix('settings')->group(function () {
+        Route::get('security', 'edit')->name('edit');
+        Route::delete('security', 'destroy')->name('destroy');
+        Route::put('security', 'deactivate')->name('deactivate');
+    });
 });
