+++ tests/Feature/Settings/ProfileUpdateTest.php
@@ -57,38 +57,4 @@ describe('Profile Update tests', function () {
         expect($user->refresh()->email_verified_at)->not->toBeNull();
     });
 
-    test('user can delete their account', function () {
-        $user = User::factory()->create();
-
-        $response = $this
-            ->actingAs($user)
-            ->delete('/settings/profile', [
-                'password' => 'password',
-            ]);
-
-        $response
-            ->assertSessionHasNoErrors()
-            ->assertRedirect('/');
-
-        $this->assertGuest();
-        expect($user->fresh())->toBeNull();
-    });
-
-    test('correct password must be provided to delete account', function () {
-        $user = User::factory()->create();
-
-        $response = $this
-            ->actingAs($user)
-            ->from('/settings/profile')
-            ->delete('/settings/profile', [
-                'password' => 'wrong-password',
-            ]);
-
-        $response
-            ->assertSessionHasErrors('password')
-            ->assertRedirect('/settings/profile');
-
-        expect($user->fresh())->not->toBeNull();
-    });
-
 });
